PR: Obtained from: Submitted by: Reviewed by:
Reverting the change of AccessControlList to an Interface. Reimplementation
as a "normal" class. DBSecurityService and LDAPService no longer use the
interface but again the class. Martin: Heads up, this may break Tambora
once again.

All problems with "incompatible class change" when moving from TDK 2.1 to
TDK 2.2 should be gone again. If there are still problems, let me know.

JvZ and I agreed to re-implement this change in a later revision of Turbine
again but not now as it seem to break previously running code.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220515 13f79535-47bb-0310-9956-ffa450edef68