Corrected problem with JavaDoc pointed out by Bruce Altner <baltner at hq.nasa.gov>:
A while back I referred to a problem I was having with inconsistent
behavior of the $page.setHttpEquiv() method when trying to redirect to
a different template. The javadocs (for the TemplatePageAttributes
class) suggest this syntax:

  setHttpEquiv("refresh", "5; http://localhost/nextpage.html")

whereas to get it to work I needed to write:

  setHttpEquiv("refresh", "5; URL=http://localhost/nextpage.html")

Now I can scrap the javascript!


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220529 13f79535-47bb-0310-9956-ffa450edef68