Patch by Paul Spencer <paulsp@apache.org>:
The following patch was prompted by work in the Jetspeed
project. Below is a list of the fixes and enhancement included.

o Fixed - TurbineGlobalCachedService now uses property
   services.GlobalCacheService.cache.check.frequency
   See Note 1

o Added - TurbineGlobalCachedService now uses property
   services.GlobalCacheService.cache.initial.size to when initializing
   heaptable.
   See Note 2

o Added - TurbineGlobalCachedService tries to refresh an
   RefreshableCachedObject that is expired and touched object in
   getObject()
   See Note 3

o Added - TurbineGlobalCachedService and GlobalCachedServices now allows
   the entire cache to be flushed with the flushCache().
   See Note 4

o Fixed - RefreshableCachedObject set the field created to the
   current time before refreshing the object.
   See Note 5

o Added - CachedObject now allows the expiration time to be set.
   See Note 5

Note 1: The value of property was not used.  Looks like a bug.

Note 2: This allows for a larger initial cache size to minimize
re-sizing during runtime.

Note 3: RefreshableCachedObject where refreshed during clearCache(),
but not during getObject().  This meant that the application must
re-add the object to cache in the event the object expired after the
last clearCache().

Note 4: The only to clear cache is to restart the application.

Note 5: Refreshable content may not be refreshed at regular
intervals. Nor will the time required to complete the refresh be
consistent.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220535 13f79535-47bb-0310-9956-ffa450edef68