Moving t2 build forward and aligning with other Turbine family builds.
Moved the build.properties to default.properties and created a
build.properties.sample which points lib.repo to the lib directory and
defines all the current jars in that directory as an example.

The build.xml was updated to conditionally compile in the webmacro classes
only if the jar is detected in the classpath at compile time.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220538 13f79535-47bb-0310-9956-ffa450edef68