Patch by Colin Chalmers.
This is a small/minor patch for the TurbineIntakeService.

The patch is necessary because if the String doesn't begin with "/" the
following is appended by ServletUtils "./"
This doesn't seem to be a problem with the tests I've done on catalina (win
version) but does cause problems on WAS (websphere) 3.5 running on Solaris
whereby a path similiar to the following is created
"/foo./WEB-INF/appData.ser". An error is given that the file cannot be
generated.

The fix had been tested on Catalina (win) & WAS 3.5 solaris.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220558 13f79535-47bb-0310-9956-ffa450edef68