Thanks to Peter Lynch for these patches to remove antiquated redirect logic.  Thanks also to Rodney Schneider for reminding us about these patches on several occasions.
Here's what Peter had to say about the patches:

"1. Removes the Turbine.java logic of redirection to establish new
sessions which, although reportedly a fix for certain obscure browser
and server combinations, also has the consequence of creating several
problems itself.

2. Removes the abstract requiresNewSession from SessionValidator.java
et all. This method is no longer needed as the only place it was used
was in the above removed code."

There's also a lengthy rationale for the changes here:
<http://www.mail-archive.com/turbine-dev%40jakarta.apache.org/msg03660.html>


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220620 13f79535-47bb-0310-9956-ffa450edef68