This is a backport from T3.0
It allows the user to query the ServletContext, ServletConfig and
the Application root directly from the Turbine class. It also
cleans the code up a little.

This is basically the first step into plugging Fulcrum into Turbine-2. I
don't want to put more of this into T2.2 before the release but as this
code is not invasive, it is nice to have it in the 2.2 release


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220719 13f79535-47bb-0310-9956-ffa450edef68