Adds parameterless init() methods to various services. By having these methods, it is possible to start the services from the Fulcrum service manager (or at least it would be possible if the services would be derived from the Fulcrum BaseService...).
This is another step for integrating Fulcrum into T2, the changes are
non-invasive and just add a very little code to the services.

This could be in the BaseService method, but then again, some of the
Services already have an init() method (e.g. FreeMarker) and this causes
the smallest upset.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@220724 13f79535-47bb-0310-9956-ffa450edef68