The getFileItem() and getFileItems() methods will now write a log message if the named parameter is not an instance of FileItem.
A new hashtable is being used for the storage of FileItems.  Previously,
FileItem instances were stored in the parameters hashtable of the superclass.
This was a mistake because that HashTable only stores String objects.  Any attempt
to call one of the getXXX() methods when the parameter was a FileItem would result
in a ClassCastException.

Thanks to Will Holcomb for pointing out this problem and suppling a patch!


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221165 13f79535-47bb-0310-9956-ffa450edef68