Misc javadocs updates
Methods which access instance variables now call assertInitialized() to verify that one of
the init() methods has been called.  If the object has not been initialized yet, an
IllegalStateException will be thrown.

All instances where an Enumeration or Iterator was declared outside the scope of the loop have
been changed to a for loop declaring the Iterator or Enumeration in the for statement.  All but
one of the loops using Enumeration have been changed to use Iterator instead.

If an exception is thrown during the execution of remove(int,String), a log message will be
written.

Deprecated calls to HttpResponse.encodeRedirectUrl() and HttpResponse.encodeUrl() have been
updated to HttpResponse.encodeRedirectURL() and HttpResponse.encodeURL() respectively.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221166 13f79535-47bb-0310-9956-ffa450edef68