This class now inherits from SimpleEmail in commons-email instead of using delegation. Again, I could see no benefit from using delegation other than additional coding.
setTo() has been deprecated.  Users should use addTo() from the superclass version instead.

setFrom(name,email) has been removed.  There is a superclass version of this method however,
the parameters are reversed.  It was not possible for me to leave this method and deprecate
it since the both parameters are strings.  This is a breaking change.

Misc style fixes and javadoc updates are included as well.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221198 13f79535-47bb-0310-9956-ffa450edef68