Made the naming scheme uniform for accessing all data types.  There is now a method for accessing the primative and object type of each one.  Like methods have also been added (where missing) for the arrays of the primative and object types.
The naming scheme:
  get<datatype> - primative
  get<datatype>Object - object
  get<datatype>s - array of primative
  get<datatype>Objects - array of objects

The get<datatype> and get<datatype>Object methods are overloaded to accept a default value.

I did not create a getBooleanObjects or getBooleans because it did not seems to make sense for
boolean types.

I did not create a getByteObjects because getBytes was not consistent with the other methods
returning arrays of primarives.  This method returns an array of bytes from a string.

The getBool, getInteger, and getIntegers methods have been deprecated in favor of new methods
with the consistent naming scheme.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221313 13f79535-47bb-0310-9956-ffa450edef68