Logging of convertion failures is now handled by calling logConvertionFailure.  This does not log the entrie stack trace any longer.
Some of the getXXX methods were rearranged slightly
to put the try/catch block around the actual statement that will through the
NumberFormatException.  This was needed to allow access to the value of the paramter for logging.

If a ClassCastException is caught inside of getString() a fatal level log message is written.
This should never happen.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221315 13f79535-47bb-0310-9956-ffa450edef68