Submitted by:	J�rgen Hoffmann Reviewed by:	Scott Eade
Fixes the intake validators so that validation halts when when the required rule has a value of false and nothing has been entered.

I also changed DefaultValidator so that a minLength > 0 does not equate to required = true - IMHO the minLength should only apply when a value is required.

Thanks J�rgen.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/branches/TURBINE_2_3_BRANCH@221683 13f79535-47bb-0310-9956-ffa450edef68