Patch thanks to Leandro Rodrigo Saad Cruz:
This is just a little enhacement. When I write webapps with turbine I
usually initialize my modules to get a handle to a ComponentManager or
any other component that I'd like to use inside that module.
IMHO, turbine could be changed to handle all modules as Components,
handling their life cycle and life style.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221860 13f79535-47bb-0310-9956-ffa450edef68