One of the really scary "how could this ever work without?" patches.
This allows the application to define the default content encoding of
incoming requests. Which in turn allows an all-UTF-8 application to
declare the POST forms to contain UTF-8 chars and keep all the special
characters alive.

Hint: If you ever wanted to chase a red herring, try this with a
RequestDumperValve in Tomcat...

Default encoding is ISO-8859-1, which is an improvement over the
Turbine 2.3.x US-ASCII, but this patch really shines in connection
with UTF-8.

Definitely more testing with other containers besides Tomcat 5.0.x
wanted! req.setCharacterEncoding() is not the most stable method in
the various web containers.

(The functionality of this patch would be a perfect match for a Valve
in the Pipeline. This valve would be the first thing in the pipeline
and set the CharacterEncoding of the incoming request).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/branches/TURBINE_2_3_BRANCH@222033 13f79535-47bb-0310-9956-ffa450edef68